Implemented road placement and deletion
The player can now place roads by clicking on the road button on the right menu. Clicking it will open a submenu that allows the user to toggle between placing and deleting roads. The cursor will change depending on the placement mode. Additionally, buildings are now required to be placed next to roads.