BugFix
Fixed bug where tree count was not being reset every time updatetreecount was called in pollution manager. In the current implementation of the game, this does not cause any bugs as updatetreecount is only called once immediately after the map is initialized. However, this can cause issues down the road and has been fixed